Rename AzureWebAppDiagnostics to Azure.AppServicesDiagnostics
