+++ src/Microsoft.Extensions.Logging.Azure.AppServicesDiagnostics/AzureWebAppDiagnosticsFactoryExtensions.cs
@@ -1,8 +1,8 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using Microsoft.Extensions.Logging.AzureWebAppDiagnostics;
-using Microsoft.Extensions.Logging.AzureWebAppDiagnostics.Internal;
+using Microsoft.Extensions.Logging.Azure.AppServicesDiagnostics;
+using Microsoft.Extensions.Logging.Azure.AppServicesDiagnostics.Internal;
 
 namespace Microsoft.Extensions.Logging
 {
