+++ src/Microsoft.Extensions.Logging.Azure.AppServicesDiagnostics/Internal/ICloudAppendBlob.cs
@@ -4,7 +4,7 @@
 using System.IO;
 using System.Threading.Tasks;
 
-namespace Microsoft.Extensions.Logging.AzureWebAppDiagnostics.Internal
+namespace Microsoft.Extensions.Logging.Azure.AppServicesDiagnostics.Internal
 {
     /// <summary>
     /// Represents an append blob, a type of blob where blocks of data are always committed to the end of the blob.
