+++ src/Microsoft.Extensions.Logging.Azure.AppServicesDiagnostics/Internal/WebAppLogConfigurationBuilder.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-namespace Microsoft.Extensions.Logging.AzureWebAppDiagnostics.Internal
+namespace Microsoft.Extensions.Logging.Azure.AppServicesDiagnostics.Internal
 {
     /// <summary>
     /// Used to create instances of <see cref="WebAppLogConfiguration"/>
