+++ src/Microsoft.Extensions.Logging.Azure.AppServicesDiagnostics/Microsoft.Extensions.Logging.Azure.AppServicesDiagnostics.xproj
