+++ src/Microsoft.Extensions.Logging.Azure.AppServicesDiagnostics/Properties/AssemblyInfo.cs
