+++ src/Microsoft.Extensions.Logging.Azure.AppServicesDiagnostics/project.json
@@ -1,5 +1,5 @@
 {
-  "version": "1.1.0-*",
+  "version": "1.0.0-*",
   "description": "Diagnostics logger for Azure WebApps",
   "packOptions": {
     "tags": [
