+++ test/Microsoft.Extensions.Logging.Azure.AppServicesDiagnostics.Test/Microsoft.Extensions.Logging.Azure.AppServicesDiagnostics.Test.xproj
@@ -7,7 +7,7 @@
   <Import Project="$(VSToolsPath)\DotNet\Microsoft.DotNet.Props" Condition="'$(VSToolsPath)' != ''" />
   <PropertyGroup Label="Globals">
     <ProjectGuid>b4a43221-de95-47bb-a2d4-2dc761fc9419</ProjectGuid>
-    <RootNamespace>Microsoft.Extensions.Logging.AzureWebApps.Test</RootNamespace>
+    <RootNamespace>Microsoft.Extensions.Logging.Azure.AppServicesDiagnostics.Test</RootNamespace>
     <BaseIntermediateOutputPath Condition="'$(BaseIntermediateOutputPath)'=='' ">.\obj</BaseIntermediateOutputPath>
     <OutputPath Condition="'$(OutputPath)'=='' ">.\bin\</OutputPath>
   </PropertyGroup>
