+++ test/Microsoft.Extensions.Logging.Azure.AppServicesDiagnostics.Test/SerilogLoggerProviderTests.cs
@@ -1,7 +1,6 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using System;
 using Microsoft.Extensions.Logging.AzureWebAppDiagnostics.Internal;
 using Moq;
 using Serilog;
@@ -9,7 +8,7 @@ using Serilog.Core;
 using Serilog.Events;
 using Xunit;
 
-namespace Microsoft.Extensions.Logging.AzureWebAppDiagnostics.Test
+namespace Microsoft.Extensions.Logging.Azure.AppServicesDiagnostics.Test
 {
     public class SerilogLoggerProviderTests
     {
