+++ test/Microsoft.Extensions.Logging.Azure.AppServicesDiagnostics.Test/SettingsFileContent.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-namespace Microsoft.Extensions.Logging.AzureWebAppDiagnostics.Test
+namespace Microsoft.Extensions.Logging.Azure.AppServicesDiagnostics.Test
 {
     // The format of this file is defined by the Azure Portal. Do not change
     internal class SettingsFileContent
