+++ test/Microsoft.Extensions.Logging.Azure.AppServicesDiagnostics.Test/project.json
